refactor: implemented custom hooks for clean code
