+++ frontend/src/app/(routes)/(overview)/overview-components/OverviewPage.tsx
@@ -1,28 +1,19 @@
 'use client';
-import React, { useCallback, useEffect, useMemo } from 'react';
+import React, { useEffect } from 'react';
 import { RootState } from '../../../..//store/store';
-import { parseBalance } from '@/utils/denom';
 import { getBalances } from '@/store/features/bank/bankSlice';
-import { getIBCBalances } from '@/utils/ibc';
-import chainDenoms from '@/utils/chainDenoms.json';
-import { useAppDispatch, useAppSelector } from '@/hooks/StateHooks';
+import { useAppDispatch, useAppSelector } from '@/custom-hooks/StateHooks';
 import {
   getDelegations,
   getAllValidators,
 } from '@/store/features/staking/stakeSlice';
-import { USD_CURRENCY } from '@/utils/contants';
-
-const chainDenomsData = chainDenoms as AssetData;
+import useGetAssetsAmount from '@/custom-hooks/useGetAssetsAmount';
+import useGetSortedChainIDs from '@/custom-hooks/useGetSortedChainIDs';
+import useGetIBCSortedChainIDs from '@/custom-hooks/useGetIBCSortedChainIds';
 
 const OverviewPage = () => {
   const dispatch = useAppDispatch();
   const networks = useAppSelector((state: RootState) => state.wallet.networks);
-  const stakingChains = useAppSelector(
-    (state: RootState) => state.staking.chains
-  );
-  const balanceChains = useAppSelector(
-    (state: RootState) => state.bank.balances
-  );
   const nameToChainIDs: Record<string, string> = useAppSelector(
     (state: RootState) => state.wallet.nameToChainIDs
   );
@@ -30,196 +21,16 @@ const OverviewPage = () => {
     (chainName) => nameToChainIDs[chainName]
   );
 
-  const convertToDollars = (denom: string, amount: number) => {
-    // todo: after common slice
-    /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
-    const tokensPriceInfo: any = {};
-    if (tokensPriceInfo?.[denom]) {
-      const price = +tokensPriceInfo?.[denom]?.info?.[USD_CURRENCY] || 0;
-      return amount * price;
-    }
-    return 0;
-  };
-
-  const getDenomInfo = useCallback(
-    (
-      chainID: string
-    ): { minimalDenom: string; decimals: number; chainName: string } => {
-      const config = networks?.[chainID]?.network?.config;
-      const currency = config?.currencies?.[0];
-      const chainName = config?.chainName.toLowerCase();
-      return {
-        minimalDenom: currency.coinMinimalDenom,
-        decimals: currency.coinDecimals || 0,
-        chainName,
-      };
-    },
-    [networks]
-  );
-
-  const calculateTotalStakedAmount = useCallback(() => {
-    let totalStakedAmount = 0;
-    chainIDs.forEach((chainID) => {
-      const staked = stakingChains?.[chainID]?.delegations?.totalStaked || 0;
-      if (staked > 0) {
-        const { minimalDenom, decimals } = getDenomInfo(chainID);
-        totalStakedAmount += convertToDollars(
-          minimalDenom,
-          staked / 10 ** decimals
-        );
-      }
-    });
-    return totalStakedAmount;
-  }, [chainIDs, stakingChains, getDenomInfo]);
-
-  // Todo: can be added after distribution slice is added
-  //   const calculateTotalPendingAmount = useCallback(() => {
-  //     let totalRewards = 0;
-  //     chainIDs.forEach((chainID) => {
-  //       const rewards =
-  //         distributionChains?.[chainID]?.delegatorRewards?.totalRewards || 0;
-  //       if (rewards > 0) {
-  //         const denom =
-  //           networks?.[chainID]?.network?.config?.currencies?.[0]
-  //             ?.coinMinimalDenom;
-  //         const decimals =
-  //           networks?.[chainID]?.network?.config?.currencies?.[0]?.coinDecimals ||
-  //           0;
-  //         totalRewards += convertToDollars(denom, rewards / 10 ** decimals);
-  //       }
-  //     });
-  //     return totalRewards;
-  //   }, [chainIDs, distributionChains, networks]);
-
-  const calculateTotalAvailableAmount = useCallback(() => {
-    let totalBalance = 0;
-    let totalIBCBalance = 0;
-
-    chainIDs.forEach((chainID) => {
-      const { minimalDenom, decimals, chainName } = getDenomInfo(chainID);
-      const balance = parseBalance(
-        balanceChains?.[chainID]?.list || [],
-        decimals,
-        minimalDenom
-      );
-
-      const ibcBalances = getIBCBalances(
-        balanceChains?.[chainID]?.list,
-        minimalDenom,
-        chainName
-      );
-
-      for (let i = 0; i < ibcBalances?.length; i++) {
-        totalIBCBalance += convertToDollars(
-          ibcBalances[i].balance.denom,
-          parseBalance(
-            [ibcBalances[i].balance],
-            ibcBalances?.[i]?.decimals,
-            ibcBalances?.[i]?.balance.denom
-          )
-        );
-      }
-
-      if (balanceChains?.[chainID]?.list?.length > 0) {
-        totalBalance += convertToDollars(minimalDenom, balance);
-      }
-    });
-
-    return { totalBalance, totalIBCBalance };
-  }, [chainIDs, balanceChains, getDenomInfo]);
-
-  const getSortedChainIds = useCallback(() => {
-    let sortedChains: { chainID: string; usdValue: number }[] = [];
-
-    chainIDs.forEach((chainID) => {
-      const { minimalDenom, decimals } = getDenomInfo(chainID);
-      const balanceAmountInDenoms = parseBalance(
-        balanceChains?.[chainID]?.list || [],
-        decimals,
-        minimalDenom
-      );
-
-      // minimalDenom
-      const stakedAmountInDenoms: number =
-        stakingChains?.[chainID]?.delegations?.totalStaked || 0;
-      // minimalDenom
-      // Todo: after distribution slice is added
-      //   const rewards =
-      //     distributionChains?.[chainID]?.delegatorRewards?.totalRewards || 0;
-      const rewardsAmountInDenoms: number = 0;
-      const chain: { chainID: string; usdValue: number } = {
-        chainID,
-        usdValue: 0,
-      };
-      if (balanceChains?.[chainID]?.list?.length > 0) {
-        chain.usdValue =
-          convertToDollars(minimalDenom, balanceAmountInDenoms) +
-          convertToDollars(
-            minimalDenom,
-            stakedAmountInDenoms / 10 ** decimals
-          ) +
-          convertToDollars(
-            minimalDenom,
-            rewardsAmountInDenoms / 10 ** decimals
-          );
-      }
-      sortedChains = [...sortedChains, chain];
-    });
-    sortedChains.sort((x, y) => y.usdValue - x.usdValue);
-    return sortedChains.map((chain) => chain.chainID);
-  }, [chainIDs, getDenomInfo, balanceChains]);
-
-  const getIBCSortedChainIds = useCallback(() => {
-    let sortedIBCChains: {
-      usdValue: number;
-      usdPrice: number;
-      balanceAmount: number;
-      chainName: string;
-      denomInfo: (IBCAsset | NativeAsset)[];
-    }[] = [];
-
-    chainIDs.forEach((chainID) => {
-      const { minimalDenom, chainName } = getDenomInfo(chainID);
-      const chainBalances = balanceChains?.[chainID]?.list || [];
-      chainBalances.forEach((balance) => {
-        const denomInfo = chainDenomsData[chainName]?.filter((denomInfo) => {
-          return denomInfo.denom === balance.denom;
-        });
-        if (balance?.denom !== minimalDenom && denomInfo?.length) {
-          //   const usdDenomPrice =
-          //     tokensPriceInfo[denomInfo[0]?.origin_denom]?.info?.[USD_CURRENCY] || 0;
-          const usdDenomPrice = 0;
-          const balanceAmount = parseBalance(
-            [balance],
-            denomInfo[0].decimals,
-            balance.denom
-          );
-          const usdDenomValue = balanceAmount * usdDenomPrice;
-          sortedIBCChains = [
-            ...sortedIBCChains,
-            {
-              usdValue: usdDenomValue,
-              usdPrice: usdDenomPrice,
-              balanceAmount: balanceAmount,
-              chainName: chainName,
-              denomInfo: denomInfo,
-            },
-          ];
-        }
-      });
-    });
-
-    sortedIBCChains.sort((x, y) => y.usdValue - x.usdValue);
-    return sortedIBCChains;
-  }, [chainIDs, balanceChains, getDenomInfo]);
+  const [totalStakedAmount, totalAvailableAmount, totalRewardsAmount] =
+    useGetAssetsAmount();
+  const [nativeSortedChainIds] = useGetSortedChainIDs();
+  const [ibcSortedChainIds] = useGetIBCSortedChainIDs();
 
   useEffect(() => {
     chainIDs.forEach((chainID) => {
       const allChainInfo = networks[chainID];
       const chainInfo = allChainInfo.network;
       const address = allChainInfo?.walletInfo?.bech32Address;
-      //   const denom =
-      //     networks?.[chainID]?.network?.config?.currencies?.[0]?.coinMinimalDenom;
       const basicChainInputs = {
         baseURL: chainInfo.config.rest,
         address,
@@ -247,42 +58,17 @@ const OverviewPage = () => {
     });
   }, []);
 
-  const {
-    totalBalance: totalAvailableAmount,
-    totalIBCBalance: totalIBCAssetsAmount,
-  } = useMemo(
-    () => calculateTotalAvailableAmount(),
-    [calculateTotalAvailableAmount]
-  );
-  const totalStakedAmount = useMemo(
-    () => calculateTotalStakedAmount(),
-    [calculateTotalStakedAmount]
-  );
-  //   const totalPendingAmount = useMemo(
-  //     () => calculateTotalPendingAmount(),
-  //     [calculateTotalPendingAmount]
-  //   );
-
-  const sortedChainIds = useMemo(
-    () => getSortedChainIds(),
-    [getSortedChainIds]
-  );
-
-  const ibcSortedChains = useMemo(
-    () => getIBCSortedChainIds(),
-    [getIBCSortedChainIds]
-  );
   return (
     <div>
-      {JSON.stringify(sortedChainIds)}
+      {JSON.stringify(nativeSortedChainIds)}
       <br />
       <br />
-      {JSON.stringify(ibcSortedChains)}
+      {JSON.stringify(ibcSortedChainIds)}
       <br />
       <br />
       {totalAvailableAmount +
         ' ' +
-        totalIBCAssetsAmount +
+        totalRewardsAmount +
         ' ' +
         totalStakedAmount}
     </div>
