+++ frontend/src/custom-hooks/useConvertToDollars.ts
@@ -1,0 +1,7 @@
+const useConvertToDollars = (denom: string, amount: number) => {
+  // Todo: after implementing common slice
+  const price = 1;
+  return amount * price;
+};
+
+export default useConvertToDollars;
