+++ frontend/src/utils/errors.ts
@@ -1,3 +1,4 @@
 export const ERR_NO_OFFLINE_AMINO_SIGNER =
   'OfflineAmino signer is required to sign amino messages';
 export const ERR_UNKNOWN = 'unknown error';
+export const ERR_MISSING_TOKEN = (denom:string) => `${denom} is not present in stored tokens. Please add the token!`;
